requirements.txt
django == 3.1.2
pillow
django-rest-framework